Analyze the test path after it being created
We added a test checking for isDirectory() a while ago and it will only
return true if the file is already on disk.

This was an issue in practice only if there was only one file and it
became a lot more apparent with the caching I will introduce in the
follow-up commit.